voting/pyproject.toml
[project]
name = "voting"
version = "0.1.0"
description = "Voting Service for DemoVox"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn>=0.29.0",
    "pydantic>=2.6.4",
    "prometheus-fastapi-instrumentator>=7.0.0",
    "pydantic-settings>=2.2.1",
    "redis>=5.0.3",
    "aiokafka>=0.10.0",
    "sqlalchemy[asyncio]>=2.0.29",
    "asyncpg>=0.29.0",
    "httpx>=0.28.1",
    "opentelemetry-instrumentation-fastapi>=0.62b0",
    "opentelemetry-exporter-otlp>=1.15.0",
    "protobuf>=5.0.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
    "testcontainers[kafka,postgres,redis]>=4.14.2",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]
pythonpath = ["."]
